% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galias.R
\name{galias}
\alias{galias}
\title{Retrieves the potential gene aliases from NCBI Gene Database}
\usage{
galias(x)
}
\arguments{
\item{x}{is the gene of interest}
}
\value{
returns a list containing the aliases of each one of the genes passed to it.
}
\description{
The function looks at NCBI Gene Database entry for the specified gene and returns the list of aliases (if available).
The function is designed to retrieve data only for Homo sapies specie.
}
