% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgenes.R
\name{hgenes}
\alias{hgenes}
\title{Retrieves genes from HMDB}
\usage{
hgenes(x)
}
\arguments{
\item{x}{is the metabolite of interest}
}
\value{
returns a list of the gene names related to a specific metabolite.
}
\description{
The function looks at HMDB entry and retrieves the genes related to specific metabolite.
#'
}
