% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggID.R
\name{hkeggid}
\alias{hkeggid}
\title{Obtains the KEGG ID corresponding to a HMDB coded metabolite}
\usage{
hkeggid(x)
}
\arguments{
\item{x}{is the HMDB coded metabolite of which we want the KEGG ID.}
}
\value{
returns the KEGG IDs of the metabolite of interest.
}
\description{
The function looks at the HMDB entry and retrieves the KEGG ID codification specific for the metabolite of interest, if available.
}
