% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsumm.R
\name{hsumm}
\alias{hsumm}
\title{Retrieve all available information of interest from HMDB entry}
\usage{
hsumm(x)
}
\arguments{
\item{x}{is the metabolite codified in HMDB ID}
}
\value{
returns a data.frame of different data entries: Acces for accession nubmer, Name for the metabolite name, Formula for its formula, SMILES for
the formula in SMILES form, Genes for the list of genes related to the metabolite and Paths for the metabolic pathways related to the metabolite.
}
\description{
Starting from HMDB ID, it retrieves the Accession number of the metabolite, the name and chemical and SMILE formula,
the diseases/pathways described in which it is involved, the KEGG ID to connect with KEGG database and the genes that are
related to it.
}
