% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenz.R
\name{kenz}
\alias{kenz}
\title{Retrieve enzymes related to specific metabolite}
\usage{
kenz(x)
}
\arguments{
\item{x}{is the metabolite we want to obtain information from}
}
\value{
returns EC code of enzymes related to the metabolite present in KEGG database
}
\description{
Retrieve enzymes related to specific metabolite
}
