% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kenzgen.R
\name{kenzgen}
\alias{kenzgen}
\title{Retrieve genes related to a specific enzyme}
\usage{
kenzgen(x)
}
\arguments{
\item{x}{is the enzyme that we want information from}
}
\value{
names of the genes related to the enzyme
}
\description{
It takes the EC code of the enzyme and searches on KEGG database for related genes
}
