% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgenes.R
\name{kgenes}
\alias{kgenes}
\title{Retrieve genes related to a specific metabolite}
\usage{
kgenes(met)
}
\arguments{
\item{met}{is the metabolite we want to obtain information from}
}
\value{
returns gene names related to the metabolite
}
\description{
Starting from a metabolite KEGG ID, it searches for the related enzymes and, then, the genes that regulate these enzymes.
}
