% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpaths.R
\name{kpaths}
\alias{kpaths}
\title{Retrieve pathways previously described in KEGG database}
\usage{
kpaths(x)
}
\arguments{
\item{x}{is the metabolite we want information from}
}
\value{
returns the pathways in which the metabolite is involved
}
\description{
Using the metabolite KEGG ID, it looks for the pathways already published in KEGG and returns them.
}
