% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksumm.R
\name{ksumm}
\alias{ksumm}
\title{Retrieve all available information of interest from KEGG entry}
\usage{
ksumm(x)
}
\arguments{
\item{x}{is the metabolite codified in KEGG ID}
}
\value{
returns a data.frame of different data entries: Enzymes, Genes and Pathways
}
\description{
Starting from KEGG ID, it retrieves the Enzymes, Genes and Pathways related to the specified metabolite
}
