% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omim.R
\name{omim}
\alias{omim}
\title{Retrieves diseases and metabolic pathways alttered present in OMIM database.}
\usage{
omim(x)
}
\arguments{
\item{x}{is the gene of interest.}
}
\value{
returns a list of the different alterations described for the gene.
}
\description{
For a gene, the functions searches on OMIM database and retrieves what is publicated there, the diseases and metabolic alterations reported.
}
