% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owl_code.R
\name{owl_code}
\alias{owl_code}
\title{Translates OWL code to KEGG and HMDB ones}
\usage{
owl_code(x)
}
\arguments{
\item{x}{is the Owl-coded metabolite of interest}
}
\value{
returns both KEGG and HMDB IDs.
}
\description{
The function looks at Owl file, which includes OWL code, metabolite name and family to obtain both the corresponding KEGG and HMDB IDs.
#'
}
