% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{pubmed}
\alias{pubmed}
\title{Obtains information from PubMed Database}
\usage{
pubmed(x, keywords_pubmed)
}
\arguments{
\item{x}{is the gene we want to look for information}

\item{keywords_pubmed}{is a character string that limits your search}
}
\value{
returns a list of ID of GEO entries, the description, title, pubmed_id and citation details of each GEO entry.
}
\description{
Using a series of keywords, this function accesses the PubMed database and look for publications related to specific gene, filtering the results depending on the keywords used.
It is generated in order to filter for publications related to Homo sapiens specie.
}
\details{
keywords_pubmed Depending on what we want, we have different options:
- to filter for words in Title/Abstract add [Title/Abtract] after the word.
- to filter for words in the general text, add [Text Word] after the word.
- to filter por publication type, add [ptyp] after the word. For example, for a review: Review[ptyp].
keywords_pubmed object is generated as follows: keywords_pubmed<-c("[Title/Abtract] whatever text we want")
}
