% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotides.R
\name{search_nucl}
\alias{search_nucl}
\title{Retrieves accession numbers of nucleotide transcripts from NCBI}
\usage{
search_nucl(x)
}
\arguments{
\item{x}{is the gene of interest}
}
\value{
returns a list of nucleotide accession numbers.
}
\description{
Using a gene name, it retrieves the transcripts available on NCBI database. The function was developed to look for specifically Homo sapiens sequences.
}
